+++ packages/solana/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./client";
 export * from "./errors";
+export * from "./events";
 export * from "./pda";
 export * from "./types";